server/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
djangorestframework==3.17.0
dotenv==0.9.9
gunicorn==25.1.0
idna==3.11
oauthlib==3.3.1
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
